requirements.txt
google-genai==0.3.0
feedparser==6.0.11
requests==2.31.0
